+++ spec/www/spec/db-open-close-delete-test.js
@@ -1935,6 +1935,8 @@ var mytests = function() {
         });
 
         test_it(suiteName + ' repeatedly open and close database faster (5x)', function () {
+          pending('TBD SKIP FOR NOW (...)'); // XXX TBD ???
+
           // TBD CURRENTLY BROKEN on iOS/macOS due to current background processing implementation:
           if (isAppleMobileOS || isMac) pending('CURRENTLY BROKEN on iOS/macOS (background processing implementation)');
           // TBD ???:
@@ -1996,6 +1998,8 @@ var mytests = function() {
 
         // (Was needed to test support for an enterprise application)
         test_it(suiteName + ' repeatedly open and delete database (4x)', function () {
+          pending('TBD SKIP FOR NOW (...)'); // XXX TBD ???
+
           var dbName = 'test-repeatedly-open-and-delete-4x.db';
           var dbargs = {name: dbName, iosDatabaseLocation: 'Documents'};