+++ src/android/io/sqlc/SQLitePlugin.java
@@ -447,6 +447,7 @@ public class SQLitePlugin extends CordovaPlugin {
         final int dbid;
         final String dbname;
         final String dblocation;
+        final String dbkey;
         // expose oldImpl:
         boolean oldImpl;
         private boolean bugWorkaround;
@@ -460,6 +461,18 @@ public class SQLitePlugin extends CordovaPlugin {
         DBRunner(final String dbname, JSONObject options, CallbackContext cbc, int dbid) {
             this.dbid = dbid;
             this.dbname = dbname;
+
+            String key = ""; // (no encryption by default)
+            if (options.has("key")) {
+                try {
+                    key = options.getString("key");
+                } catch (JSONException e) {
+                    // NOTE: this should not happen!
+                    Log.e(SQLitePlugin.class.getSimpleName(), "unexpected JSON error getting password key, ignored", e);
+                }
+            }
+            this.dbkey = key;
+
             this.oldImpl = options.has("androidOldDatabaseImplementation");
             //Log.v(SQLitePlugin.class.getSimpleName(), "Android db implementation: built-in android.database.sqlite package");
             this.bugWorkaround = this.oldImpl && options.has("androidBugWorkaround");
@@ -488,6 +501,9 @@ public class SQLitePlugin extends CordovaPlugin {
                     this.mydb = this.mydb1 = openDatabase(dbname, dblocation, this.openCbc, this.oldImpl, this.dbid);
                 else
                     this.mydb = openDatabase2(dbname, dblocation, this.openCbc, this.oldImpl);
+
+                if (!oldImpl && this.dbkey != "")
+                    EVPlusNativeDriver.sqlc_db_key_native_string(mydb1.mydbhandle, this.dbkey);
             } catch (Exception e) {
                 Log.e(SQLitePlugin.class.getSimpleName(), "unexpected error, stopping db thread", e);
                 dbrmap.remove(dbname);