+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Changes
 
+## cordova-sqlcipher-evplus-ext-build-free 0.0.1-dev
+
+- SQLCipher support with evplus enhancements for Android
+
 ## cordova-sqlite-evplus-feat-android-db-location 0.0.1-dev
 
 ### cordova-sqlite-evcore-feature-android-db-location 0.7.2-dev1